# backend/componentsdb/api/models.py
# generated by datamodel-codegen:
#   filename:  spec.yaml

from __future__ import annotations

from typing import Annotated, List, Optional
from uuid import UUID

from pydantic import BaseModel, Field


class ServerStatus(BaseModel):
    pass


class DrawerSummary(BaseModel):
    id: UUID
    label: str


class DrawerDetail(DrawerSummary):
    pass


class CabinetSummary(BaseModel):
    id: UUID
    name: str
    drawers: List[DrawerSummary]


class CabinetDetail(CabinetSummary):
    drawers: List[DrawerDetail]


class CabinetList(BaseModel):
    items: List[CabinetSummary]
    next_cursor: Annotated[Optional[UUID], Field(None, alias="nextCursor")]