# openapi/spec.yaml
openapi: 3.0.3
info:
  title: Components Database
  version: 0.1.0
tags:
  - name: cabinet
    description: Cabinets holding drawers of components
  - name: drawer
    description: Drawers of components located within cabinets
  - name: component
    description: Components
  - name: status
    description: API status
paths:
  /cabinets:
    get:
      tags:
        - cabinet
      summary: List cabinets
      description: List cabinets with pagination
      operationId: cabinetsList
      parameters:
        - name: cursor
          in: query
          description: Cursor used for paginated responses
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CabinetList"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /cabinets/{cabinetId}:
    get:
      tags:
        - cabinet
      summary: Get cabinet and contents
      description: Get contents of a single cabinet
      operationId: cabinetGet
      parameters:
        - name: cabinetId
          in: path
          description: ID of cabinet to return
          schema:
            type: string
            format: uuid
          required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CabinetDetail"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /status:
    get:
      tags:
        - status
      operationId: statusGet
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStatus"
components:
  schemas:
    CabinetSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        drawers:
          type: array
          items:
            $ref: "#/components/schemas/DrawerSummary"
      required: ["id", "name", "drawers"]
    ServerStatus:
      type: object
    CabinetDetail:
      type: object
      allOf:
        - $ref: "#/components/schemas/CabinetSummary"
        - type: object
          properties:
            drawers:
              type: array
              items:
                $ref: "#/components/schemas/DrawerDetail"
          required: ["drawers"]
    CabinetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CabinetSummary"
        nextCursor:
          type: string
          format: uuid
      required: ["items"]
    DrawerSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
      required: ["id", "label"]
    DrawerDetail:
      type: object
      allOf:
        - $ref: "#/components/schemas/DrawerSummary"